Abbreviate Plan months and restore 44px week controls.
The Plan range label used full month names, so "31 August – 6 September"
needed 295px of a 358px phone header and wrapped under the week controls.
Shortening the month to "31 Aug – 6 Sept" cuts it to 189px.

Spend the reclaimed width on touch targets. The arrows and the This week
chip go back to the stock 44px icon size, and the sizing breakpoint moves
from a bespoke sm: to the md: the button variants already use.

Three 44px controls fit beside the longest common label; four do not. The
header Add meal is therefore hidden below md, where the floating button,
every empty meal slot, and the library empty state all still offer it.

+++ src/lib/read-models/plan.ts
@@ -70,7 +70,7 @@ const dayFormatter = new Intl.DateTimeFormat("en-GB", {
   timeZone: ZURICH_TIME_ZONE,
 });
 const monthFormatter = new Intl.DateTimeFormat("en-GB", {
-  month: "long",
+  month: "short",
   timeZone: ZURICH_TIME_ZONE,
 });
 const yearFormatter = new Intl.DateTimeFormat("en-GB", {
